openclaw/extensions/trading-engine/python/requirements.txt
finrl>=0.3.6
stable-baselines3>=2.3.0
gymnasium>=0.29.0
# torch is installed separately in Dockerfile (CPU-only, saves ~4GB)
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=9.0,<11
pydantic>=2.9.0
httpx>=0.27.0
hmmlearn>=0.3.0
scikit-learn>=1.5.0
xgboost>=2.0.0
transformers>=4.40.0
alpaca-trade-api>=3.0.0
ccxt>=4.5.0
mojito2>=0.2.0
apscheduler>=3.10.0
fastapi-mcp>=0.1.0
claude-agent-sdk>=0.1.19
requests>=2.31.0
pytest>=8.0.0
